added new image utils method to update the image name as well (#276)
## Description

While developing the new system tests for the Kroxylicious operator, I
realized that if I wanted to change also the name if the image of the
repository, it was not possible because the current method didn't change
that.

This would be very useful for our tests because we need to change the
name of the image for PR tests.


## Type of Change

* New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit/integration tests pass locally with my
changes

---------

Signed-off-by: Francisco Vila <fvila@redhat.com>
Co-authored-by: Lukáš Král <53821852+im-konge@users.noreply.github.com>