Add possibility to collect logs from previous (failed) Pod containers (#292)
## Description

This PR adds possibility to log previous logs of the failed Pod and its
container(s).
That is useful in cases that we are running the automatic log collection
during test-case execution - as most probably logs that are collected
are actually from the startup of the Pod and its container(s) - which
doesn't tell us much about the actual error why the Pods are failing to
get to the running state.

## Type of Change

* New feature (non-breaking change which adds functionality)

## Checklist

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit/integration tests pass locally with my
changes

---------

Signed-off-by: Lukas Kral <lukywill16@gmail.com>