LogCollector - run log collection asynchronously (#293)
## Description

Currently, when we are collecting logs and resource descriptions from
the namespace(s), we are doing this synchronously.
That means, when there are a lot of resources, it takes a really long
time until everything is collected.
This PR changes the collection to be asynchronous - so it will be
faster.

## Type of Change

* New feature (non-breaking change which adds functionality)

## Checklist

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit/integration tests pass locally with my
changes

Signed-off-by: Lukas Kral <lukywill16@gmail.com>