Add possibility to specify log level in exec commands (#295)
## Description

In one of the previous PRs the option to enable/disable logging to
output when the `exec` command is executed.
But in some cases - like commands generating long output - we want to
log it, but on different level - for example just for debug in case that
there is some failure, so we have more logs to check.

This PR adds possibility to specify log level in the `exec` commands.
For better usage (without need of adding the `slf4j` library directly
into the code) I created the `LogLevel` enum, which contains all
possible log levels and then it returns the appropriate `Level` enum
from sl4j, so it can be used in the `LOGGER.atLevel()` method.

## Type of Change

* New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit/integration tests pass locally with my
changes

---------

Signed-off-by: Lukas Kral <lukywill16@gmail.com>