Add method for mapping generic resource to specific resource and make getter in `waitResourceCondition`modifiable (#349)
## Description

This PR adds method (and new class) for handling the mapping from
`GenericKubernetesResource` to specific resource provided by user. This
mapping is useful in case that you have the object in your suite, but
it's "configured" to newer version of the API, however you would like to
handle resources with the older API version.
Then typical client for that resource would not work, you need to use
the `genericKubernetesResources` method from the Kube client and map it
to the object.

Also, this PR makes the getter for the resource in the
`waitResourceCondition` modifiable - in order to for example pass this
`ResourceUtils#getGenericResourceReturnSpecific` instead of the usual
`kubeClient().getClient().resource(resource).get()`.

## Type of Change

* New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit/integration tests pass locally with my
changes

---------

Signed-off-by: Lukas Kral <lukywill16@gmail.com>