Change handling of futures in KubeResourceManager (#351)
## Description

Currently, we are experiencing in some pipelines situation when the
tests are stuck on deletion of resources - it looks like there is not
correctly set timeout for the CompletableFutures we are using in
`deleteResources` and other methods.
That causes issues with threads that are stuck on waiting for some
resource to be deleted.

This PR changes the way how we are using the CompletableFutures -
instead of `join()` we will use `get()` with configured timeout for each
operation, which should help with time-outs during the resource deletion
- and also exception handling. Other than that, I'm adding the executor
service as it's used in `Wait` class - to correctly handle the threads.
With all these enhancements it seems that pipelines are not stuck
anymore.

## Type of Change

* Bug fix (non-breaking change which fixes an issue)

## Checklist

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit/integration tests pass locally with my
changes

---------

Signed-off-by: Lukas Kral <lukywill16@gmail.com>